+++ components/accordion/script.js
@@ -1,5 +1,3 @@
-'use strict';
-
 require('jquery-accessible-accordion-aria/jquery-accessible-accordion-aria.js');
 
 $(function () {