+++ functions.php
@@ -32,8 +32,15 @@ require get_template_directory() . '/vendor/autoload.php';
 // Création d'une instance de Pug pour le rendu des templates *.pug
 // TODO: configurer le cache, voir https://github.com/pug-php/pug
 $pug = new \Pug\Pug([
-	'expressionLanguage' => 'js'
+	'pugjs' => true,
+	// 'expressionLanguage' => 'js',
+	'basedir' => get_template_directory() . '/modules',
+	'cache' => get_template_directory() . '/cache/pug',
+	// 'upToDateCheck' => false,
 ]);
+// list($success, $errors) = $pug->cacheDirectory(get_template_directory() . '/modules');
+// echo "$success files have been cached\n";
+// echo "$errors errors occurred\n";
 
 // Fonctions utiles
 require get_template_directory() . '/inc/utile.php';
@@ -76,13 +83,11 @@ require get_template_directory() . '/inc/profile-subpages.php';
 // Redirection des non-admins vers la page d'accueil lors du login
 require get_template_directory() . '/inc/redirect-after-login.php';
 
-// Customisation du plugin Algolia
-if ( class_exists( 'Algolia_Plugin' ) ) {
-	#require get_template_directory() . '/inc/algolia.php';
-}
-
 // Customisation ACF
 require get_template_directory() . '/inc/acf.php';
+
+// Intégration d'Algolia
+require get_template_directory() . '/algolia/functions.php';
 
 
 if ( ! function_exists( 'telabotanica_setup' ) ) :
@@ -233,3 +238,14 @@ function telabotanica_admin_theme_style() {
 	wp_enqueue_style('telabotanica', get_template_directory_uri() . '/admin/style.css');
 }
 add_action('admin_enqueue_scripts', 'telabotanica_admin_theme_style');
+
+/**
+ * Adds "upload_files" capability to "contributor" role after theme switch
+ *
+ * @since TelaBotanica 0.1
+ */
+function telabotanica_add_upload_capability_to_contributor_role() {
+	$contributor = get_role('contributor');
+	$contributor->add_cap('upload_files');
+}
+add_action( 'after_switch_theme', 'telabotanica_add_upload_capability_to_contributor_role' );