+++ modules/block-dashboard-images/script.js
@@ -1,11 +1,11 @@
-'use strict';
-
 var PubSub = require('pubsub-js');
+
 var Tela = window.Tela || {};
+Tela.modules = Tela.modules || {};
 
-Tela.blockDashboardImages = (function(){
+Tela.modules.blockDashboardImages = (function(){
 
-	function blockDashboardImages(selector){
+	function module(selector){
 		var $el = $(selector),
 			$titleSuffix,
 			$content,
@@ -30,23 +30,19 @@ Tela.blockDashboardImages = (function(){
 		}
 
 		function loadData(){
+			// useful for local debugging:
+			// apiUrl = '/wp-content/themes/telabotanica/modules/feed/images.json';
 
 			// Call the API
-			$.ajax({
-				type: "GET",
-				url: apiUrl,
-				dataType: "xml",
-				success: function(xml){
-					$(xml).find('entry').slice(0, maxItems).each(function(){
-						var $this = $(this);
-						data.items.push({
-							href: $this.find('link').attr('href'),
-							image: $this.find('id').text().replace('L.', 'CRXS.')
-						});
+			$.getJSON(apiUrl, function(json){
+				_.each(json.resultats.slice(0, maxItems), function (item) {
+					data.items.push({
+						href: '#',
+						image: item['binaire.href']
 					});
+				});
 
-					renderContent();
-				}
+				renderContent();
 			});
 		}
 
@@ -77,12 +73,12 @@ Tela.blockDashboardImages = (function(){
 
 	return function(selector){
 		return $(selector).each(function(){
-			blockDashboardImages(this);
+			module(this);
 		});
 	};
 
 })();
 
 $(document).ready(function(){
-	Tela.blockDashboardImages('.block-dashboard-images');
+	Tela.modules.blockDashboardImages('.block-dashboard-images');
 });