+++ modules/search-box/search-box.php
@@ -2,11 +2,8 @@
 	$defaults = [
 		'id' => false,
 		'autocomplete' => true,
-		'action' => home_url( '/' ),
-		'input_id' => false,
-		'input_name' => 's',
 		'placeholder' => __('Rechercher une plante, un projet, un mot clé...', 'telabotanica'),
-		'value' => get_search_query(),
+		'value' => get_search_query() ?: get_query_var( 'q', false ),
 		'index' => false,
 		'suggestions' => false,
 		'modifiers' => ['large']
@@ -16,26 +13,25 @@
 	$data->modifiers = telabotanica_styleguide_modifiers_array('search-box', $data->modifiers);
 
 	printf(
-		'<div class="%s" id="%s" data-autocomplete="%s">',
+		'<div class="%s" data-autocomplete="%s" data-instantsearch="%s" data-index="%s">',
 		implode(' ', $data->modifiers),
-		esc_attr($data->id),
-		var_export($data->autocomplete, true)
+		var_export($data->autocomplete, true),
+		var_export($data->instantsearch, true),
+		$data->index
 	);
 		printf(
 			'<form role="search" method="get" action="%s" class="search-box-wrapper">',
 			esc_url( $data->action )
 		);
 			printf(
-				'<input name="%s" id="%s" type="text" class="search-box-input" placeholder="%s" value="%s" autocomplete="off" spellcheck="false" />',
-				esc_attr($data->input_name),
-				esc_attr($data->input_id),
-				esc_attr($data->placeholder),
-				esc_attr($data->value)
+				'<input name="s" type="text" class="search-box-input" placeholder="%s" value="%s" autocomplete="off" spellcheck="false" />',
+				esc_attr( $data->placeholder ),
+				esc_attr( $data->value )
 			);
 			if ($data->index) :
 				printf(
-					'<input name="index" type="hidden" value="%s" />',
-					esc_attr($data->index)
+					'<input name="in" type="hidden" value="%s" />',
+					esc_attr( sanitize_key( $data->index ) )
 				);
 			endif;
 			printf(